+++ changelog.md
@@ -7,6 +7,9 @@
 
 ## Change log
 
+### v2019.4.32
+FEATURE: The link chooser was missing the cancel button.
+
 ### v2019.4.31
 CHORE: Updated all dependencies.
 
@@ -60,7 +63,7 @@ FEATURE: for all links in P3X OneNote is left for the user to decide how the lin
 
 ---
 
-[**P3X-ONENOTE**](https://pages.corifeus.com/onenote) Build v2019.4.31 
+[**P3X-ONENOTE**](https://pages.corifeus.com/onenote) Build v2019.4.32 
 
 [![Donate for Corifeus / P3X](https://img.shields.io/badge/Donate-Corifeus-003087.svg)](https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=QZVM4V6HVZJW6)  [![Contact Corifeus / P3X](https://img.shields.io/badge/Contact-P3X-ff9900.svg)](https://www.patrikx3.com/en/front/contact) [![Like Corifeus @ Facebook](https://img.shields.io/badge/LIKE-Corifeus-3b5998.svg)](https://www.facebook.com/corifeus.software)